[XXXXX] Populate dataId in dataObjInfo_t on create
This change populates dataObjInfo_t.dataId on create. The information
is available at the time of registration in the catalog but rsPhyPathReg
does not return it to the caller. So, creating a new replica has been
split into two paths: new replica for existing data object (which uses
rsRegReplica) and new replica for new data object (rsRegDataObj).

Also rearranges rsPhyPathReg and boxes in some unused functions in the .hpp.