[_3848] Add replica_state_table Singleton
Adds a global Singleton a la replica_access_table, but is specific to
individual agents. It is like an L1 descriptor which keeps track of all
information for each replica of a given data object. It consists of a
map which keys on logical paths and has a value consisting of the state
of the data object in the catalog at time of open (known as the "before"
state) and an updating state which will serve as the input to a call to
data_object_finalize (known as the "after" state). Each state is
represented as a data_object_proxy which serves as a proxy to the
DataObjInfo linked lists from the legacy C API.

The "before" state exists in order to keep track of the replica status,
size, checksum, size, and other information before any changes were made
due to data movement. The "after" state exists to serve as input to a
catalog update on the data object once the operation has completed.

The replica_state_table keeps its own copies of the DataObjInfo
structures so that it can be responsible for its own memory in the event
that the client does not close the data object/replica after finishing
the operation. get_data_object_state returns a copy of an entry in the
replica_state_table, so the caller is responsible for the memory.
Likewise, set_data_object_state free's the existing entry being set and
makes a new copy of the provided data_object_proxy. In the future, this
could possibly be made to reduce the number of copies, but the C APIs
using this do not currently have a consistent way of dealing with their
own data structures.

Currently only supports one open per agent. If an operation needs
additional replicas to be opened at the same time (e.g. replication and
phymv), the operation should keep track of its own information as
needed.