[XXXXX] Add duplicate_replica
Adds a duplicate_replica method to the replica namespace in
replica_proxy.

Also fixes duplicate_data_object. DataObjInfo has a few sub-structs
which need to be allocated and copied for each duplicated node in the
linked list. Some items are not in use and are not touched by the proxy
object.

[XXXXX] Add get_replica_size_from_storage to replica lib

add data_id impl for make_data_object_proxy

add test for data_id

fix to replica_proxy