[_5478] Add multithreaded tests: irepl/iphymv/icp
Adds a test for each suite which puts files of varying sizes and asserts
that the number of threads used in the data transfer is reported as
being at or above a certain value.

There is a test for the following sizes in irepl, iphymv, and icp:
    - 0-length
    - At the threshold of the default parallel transfer minimum size
    - Above the default parallel transfer minimum size

The test_irepl suite has been rearranged such that the unittest
framework setUp() and tearDown() methods can be used for more
effectively for some tests in the suite.