[_5478] Set numThreads in rsDataObjCopy/Repl/Phymv
The number of threads being used in data transfers for copies,
replication, and physical moves needs to be calculated and set in order
to get multi-threaded transfers. The TransferStat parameter of each
server-side API needs to be informed of this information as well so that
the statistics can be reported back to the user.