[ 6385] Add DB plugin op for data_object_finalize
This adds a database plugin operation for use in the
data_object_finalize API plugin. This is meant to act as interchangeable
logic with the nanodbc-based database interactions which exist in the
API plugin today.

This change is needed in order for zone administrators to implement
policy around data objects being created or modified in the system. In
the past, this was possible because of the "mod data obj meta" database
operation. This was discarded in favor of the nanodbc-based operations
built directly into the data_object_finalize API plugin. This new
database operation should act as a replacement.

Due to logical locking, this database operation will be enacted once
when the data object is opened and once when it is closed. The rule
implementer is responsible for taking the appropriate action.