[_5759] Prevent query in file_modified
Uses the information stored in the replica list provided to
data_object_finalize to create the irods::file_object required by the
fileModified interface. This information theoretically represents the
latest information in the catalog because data_object_finalize is
supposed to atomically update the rows in the database with the
information stored in the JSON input.

A new parameter is now supported in the input for data_object_finalize.
It is optional and allows the caller to provide a logical path which
will avoid a hit to the database to resolve the coll_id provided in the
R_DATA_MAIN input. If it is not provided, a query will be run to fill
in the missing information needed to invoke fileModified.

This adds a couple of new factory methods for data_object_proxy and
file_object.