[_5842] Make the test hostnames configurable
Adds a new option to scripts/run_tests.py called --hostnames which
allows the user to specify an ordered list of hostnames by which the
test framework can identify different running iRODS servers. The option
is a space-delimited list of exactly 4 strings in this order:

    --hostnames ICAT_HOSTNAME HOSTNAME_1 HOSTNAME_2 HOSTNAME_3

...where each string maps directly to the variables of the same name in
irods.test.settings.

If --hostnames is not specified and --topology is, the values will be:
    ICAT_HOSTNAME       icat.example.org
    HOSTNAME_1          resource1.example.org
    HOSTNAME_2          resource2.example.org
    HOSTNAME_3          resource3.example.org

This is the historical behavior of run_tests.py. If --topology is not
specified, the tests will run as usual with socket.gethostname() as the
hostname for all 4 hostnames (that is, all tests are running on a single
host and it is the only host of which the tests are aware).