[_5838] New redirection uses server structures
The irods::catalog_utilities::redirect_to_catalog_provider facility
provides a cleaner interface for establishing a server-to-server
connection to the catalog service provider. This was introduced recently
and used in several of the new API plugins in the core server. It does
not add itself to the list of servers to which the agent has connected
over its lifetime which is tracked in the core server code and then
cleaned up on agent teardown right after closing all of the open L1
descriptors.

This changes irods::catalog_utilities::redirect_to_catalog_provider such
that it uses the appropriate server facilities for automatic connection
cleanup on agent teardown. In the future there may be a better way to do
this but for now the utility should conform to the existing framework.