[_5548] Stale sibling replicas in rsDataObjClose
Sibling replicas should be staled in rsDataObjClose like it is done in
replica_close. This is because newly created replicas which are not part
of the rsDataObjRepl operation must be assumed to be different from the
other replicas and therefore the other replicas should be staled when
finalizing the data object in rsDataObjClose.