[_5848] Add test for opening the same replica twice
This test opens the same replica with two different agents via a replica
token. This should be allowed and is tested by having another replica
which exists and will be write-locked when the first replica is opened.
The second open call should succeed because it should be targeting the
intermediate replica with the replica token it has.