[_4649] Prevent unpack_struct leak
A pointer to the bytesBuf_t.buf of the unpacked structures is being
assigned to the pointer to the output structure. This causes a leak in
certain usages where the structure is replicated to and from
microservice parameters.

This should be investigated to understand the true cause of the leak
further up the call stack.