[_6100] Copy hierarchies to condInput in repl/phymv
In 4.2.9, repl and phymv were rewritten in such a way that the
DataObjInp provided by the caller of these APIs is treated more as a
pure input, with copies being made and then manipulated and freed for
the source and destination replicas. This alters the historical behavior
of the API in such a way that breaks at least one Consortium-released
plugin which depends on the resource hierarchy of the destination
replica being supplied in the condInput of the DataObjInp in the
post-PEP for the repl API invocation.

This change copies the resolved hierarchies of both the source and
destination replicas back into the provided DataObjInp.condInput so that
the information is available to any client or policy which may have
depended on it in the past.