[_5479] Add free function to determine auth method
Adds a new function to the authentication plugin framework header which
allows a client or server to determine which authentication method it
ought to be using in a unified fashion. The implementation at this time
is, of course, a simple version check, but it could be based on other
criteria which can be queried from an RcComm.