[ 5479] Save password in native auth plugin
This change also fixes support for argument-supplied password from iinit
in the native authentication flow. The password has to be removed and
restored at various points in the flow to ensure that the password is
not sent in the clear over the wire to the server.

Also removes some useless messages printed when getting the client
environment.