BIG oopsie in replica_close
When an error occurs, replica_close will attempt to stale the replica
using the same mechanism as the replica status updating logic in
non-error cases. This updates all sibling replicas to STALE even though
in the error cases we only want the target replica to be stale.

This exposes the "unlock statuses" for all locked sibling replicas in
the unlock mechanism such that the error case can restore the replica
status rather than setting it to stale each time. The default behavior
is still to stale the sibling replicas as this is the most common case.