[ 6101] Add tests for GenQuery with DATA_RESC_HIER
Adds a set of tests which query for DATA_RESC_HIER with various GenQuery
operations and using wildcard characters (or not). DATA_RESC_HIER is a
special GenQuery column because DATA_RESC_HIER is not populated in the
database. Instead, GenQuery has to translate the possible results to a
set of DATA_IDs and run the query against that.