[ 6068][ 6365] Improve ils error handling tests
This change makes a couple of test adjustments:

1. Add misconfigured SSL federation test

When SSL is misconfigured in a federated environment, some iCommands
terminate with uncaught exceptions. This test ensures that the
exceptions are being caught in this situation.

2. Adjust expected error from ils in test

Now that ils catches exceptions, a couple of tests needed to be adjusted
to expect the new output instead of the uncaught exception/stacktrace
which occurred before.