[ 6267][ 6740] Check for nullptr in GenQuery inputs
Throughout GenQuery and its constellation of utility functions, there
are few or no checks on the input structures and their members to see
whether pointers are valid before derefencing and free'ing them.

This change adds a number of checks to these functions to ensure that
the inputs are valid before attempting to use them.