[ 6842] Use zone hints to redirect in rsModColl
The redirection logic for modifying collection information needs to
consider cross-federation connections. In the special case where a
redirect is occurring inside a zone for a privileged client, a new
connection needs to be made as the service account rodsadmin so that the
privileged client status will be upheld through the redirect. In the
case of a cross-federation call, the privileged client user in the local
zone may not exist or be privileged in the remote zone, so a regular
redirect is used in that case.