[ XXXX] Add test/mechanism for resource operation failure
This adds some tests which ensure that the system is behaving properly
when certain resource plugin operations fail. Specifically, when a data
object is created and some resource plugin operation fails, the data
object should be marked stale to ensure that the state of the data is
properly reflected in the catalog.

This change adds a new keyword recognized by the unixfilesystem storage
resource plugin: munge_operations. This key takes a comma-separated list
of strings which should match to resource plugin operation names. This
will be useful in testing so that we can purposely cause certain
resource operations to fail just by changing the context string on a
particular resource.