[ 6154] replica_close: Fix setting replica status
When a failure occurs during replica_close, the replica status is
updated to stale in the replica state table. However, if the L1
descriptor is not freed, the replica status will be updated again when
the agent is torn down on disconnect. This change ensures that the
replica status in the L1 descriptor table matches the intended status
(stale) so that the tear down logic does not erroneously set it back to
an intermediate or locked status.

When an error occurs, replica_close will attempt to stale the replica
using the same mechanism as the replica status updating logic in
non-error cases. This updates all sibling replicas to STALE even though
in the error cases we only want the target replica to be stale.

This exposes the "unlock statuses" for all locked sibling replicas in
the unlock mechanism such that the error case can restore the replica
status rather than setting it to stale each time. The default behavior
is still to restore the statuses for the sibling replicas as this is the
most common case.