squash with 'Always clean up and free rsComm members'
This change erases the current PID from the replica access table as part
of the agent teardown logic. If the PID was already erased in a close
operation, then nothing will be done. This is to ensure that the replica
access table only contains active PIDs. A unit test has been added for
erasing a PID which does not exist in the replica access table.