[ 6896] Add -n tests
This adds a new test suite which focuses on testing operations which
target a specific replica using the REPL_NUM_KW (which is often exposed
as the -n option for the iCommands). Weighted passthru resources can be
configured to make read and write operations vote very low or even 0,
but if a replica exists under that passthru and it has been "requested"
through this keyword for the operation, that replica is what should be
returned. If the vote for that replica is 0, it is an error.

The following iCommands got tests for these scenarios:
 - iget
 - iput
 - irepl
 - iphymv
 - istream

itouch does not open any replicas, so it does not participate in
hierarchy resolution. These tests are being skipped for now, though they
have been written.

iunreg and itrim are not being tested at this time because the unlink
operation is not available for consideration in hierarchy resolution at
this time.