squash with 'Requested replica always wins the vote'
This change also includes a peculiar workaround which is required in
order for voting to behave in the compound resource when it is
configured to prefer the archive resource (i.e. configured to "always"
overpave the cache replica). The replica in the cache resource must
contain a vote value if it is requested or it will cause operations
which request that replica by number to fail. Since the cache resource
only votes when the archive votes 0, this can happen frequently. As
such, the replica in the cache resource is now set to vote with the same
value as the replica in the archive resource (in the normal case)
because the replica in the archive resource is going to be overpaving
the replica in the cache resource anyway.