[ 6827] Search all catalog_provider_hosts in federation stanza
The federation stanza of the server_config has a property called
catalog_provider_hosts which acts as a list of catalog provider hosts
to which a local server can connect to perform operations across a
federation boundary. This change ensures that determining the
negotiation key is considering all the entries in the list instead of
just the first one.