[ 6912] Add tests for GenQuery || and DATA_RESC_HIER
Adds tests which construct a query based on the supported GenQuery
operations: LIKE, NOT LIKE, =, and !=

These are paired in combination with a condition which will yield one of
the following sets of results based on a single data object with two
replicas with distinct resource hierarchies: none, both, hierarchy 1, or
hierarchy 2

This generates a table of 256 cases. Not all cases are valid. The =
operator cannot match both hierarchies because it must be exact. The !=
operator, similarly, cannot match neither hierarchy (none) because it is
precisely NOT something (which means that it must be matching the
other).