[ 6912] GenQuery: Add support for || with DATA_RESC_HIER
This change adds support for using the || operator with DATA_RESC_HIER
in GenQuery. This operates by finding the || in the string provided by
the query input and translating the derived substring to a RESC_ID
condition. The generated clause also uses the || operator, but it is
just a series of IN (...) conditions.