[ 6997] Stage-to-cache is performed by admin user
This change temporarily elevates privileges for the client user when
performing stage-to-cache replication in the compound resource. The
client user is also switched to be the local service account rodsadmin
so that the elevated privileges will hold through redirections.

In order for scoped client identity to persist through a redirect, a new
keyword has been added which is intended for use in the RsComm
session_props key-value pair member: USE_ONEOFF_CONNECTIONS_KW.

When this keyword is present in the session_props, this signals to APIs
that redirections should use a "one-off" connection. That is, if a
connection exists between this server agent and a remote server already,
it should be disconnected and a new connection should be established.
Additionally, the new server-to-server connection should be disconnected
when the operation has completed. This is enforced by convention only.

The data object finalize API has been updated to honor the presence of
USE_ONEOFF_CONNECTIONS_KW by resetting the existing connection with the
catalog provider and disconnecting the new connection after the catalog
updates have been applied.