[ 6997] Add tests accessing archive-only objects
Adds a test for attempting to access an object which only has a replica
in an archive resource as a user with no access permission. Ensures that
stage-to-cache cannot be triggered by a user with no permissions.

Also adds a test for ibun -x on a tar file which only exists in a remote
archive resource. This ensures that the stage-to-cache required to
extract the contents can occur successfully as well as the extraction
which follows.