[ 6118] msiDataObjChksum: Add irodsAdmin support
This change allows irodsAdmin keyword to be used with msiDataObjChksum.

The use of this flag when executing the rule code as any user other than
a rodsadmin will result in a CAT_INSUFFICIENT_PRIVILEGE_LEVEL error. So,
when writing rules, the policy implementer should make sure that the
keyword is necessary and that it is allowed to be used.