[ 7098] pam_password: Remove password_length config
The pam_password authentication plugin has 2 problems:

 1. PAM passwords cannot exceed 50 characters.
 2. The password_length configuration of the pam_password authentication
    plugin_configuration has no effect.

This change removes the static buffer for holding the password which is
then sent to the server via the authenticate API. The password is now
stored in full in the JSON payload.

The password_length configuration which was previously used to govern
the size of the randomly generated password for users authenticating
with pam_password has been removed. The randomly generated password will
now be a static length of 50 characters.

The PAM password itself will not be limited in any way by iRODS and will
simply be passed along to the PAM service for authentication.

Finally, the randomly generated password for PAM authentication users
can now have an apostrophe when in scrambled form. This was prevented in
the past for compatibility with platforms for which this presented a
problem. Apostrophes are inserted into the catalog for many parts of
iRODS, so this effort is no longer necessary.