[ 7274] Use grid configurations for auth configs
Authentication configurations in the database now use grid configurations
instead of server_config. These are stored in the database and so incur
an additional query whenever they are needed. As such, the auth configs
are static and have been given the ability to mark themselves as initialized
in order to avoid duplicating the queries.