[ 7098] Fix random passwords for PAM authentication
The randomly generated passwords used with PAM authentication are not guaranteed to
be of a certain size, including having any contents at all whatsoever. This is due to
the fact that the function used to randomly generate bytes for the string may generate
all characters which fall outside the acceptable range. Furthermore, the acceptable
range of characters does not include all alphanumeric characters, but only the set
[1-8B-Yb-y].

This change uses a new function which properly generates a string of alphanumeric
characters of the specified length. Due to restrictions in password length for native
authentication, the password cannot exceed MAX_PASSWORD_LEN - 8 in length.