[ 7274] Add configuration tests for native auth
These tests are based entirely off of the pam_password equivalents
because they should behave in the same way.

Tests for password_extend_lifetime have been written but are skipped
because that configuration is not supported for native authentication.

Tests for password expiration are also being skipped because the
minimum TTL we can specify is 1 hour which is not feasible for
automated testing.