[ 6497] Add tests for icp -f -R
Adds tests which create a data object with replicas on two resources
and then attempts to overwrite the data object with icp -f while
targeting a third resource which has no replicas for the data object.

Also adds some icp tests to test_federation to ensure that icp overwrites
data objects appropriately when target resources have replicas of the data
object to overwrite.

Also adds a couple of icp tests for overwriting in other scenarios, such as
not using the force flag or trying to copy an object to overwrite itself.