[ 6497] Require icp -f -R to target resource with replica
When overwriting a data object using icp -f while targeting a specific
resource, the targeted resource ought to already have a replica to
overwrite. This brings icp in line with iput -f which returns an error
when the operation would result in a new replica on an existing data
object. icp also operates at the object (not replica) level, so the same
rules have been put in place.

The logic for determining whether the user is requesting this situation
was lifted from the put API.