[ 7476] Prefer good replicas for write voting
This commit changes the "preferred" replica status for votes on write
operations from stale to good.

One consequence of this decision is allowing replication operations to
to allow targeting good replicas for update. Instead of actually overwriting
the data in the good replica, fileModified is triggered directly to invoke
any policy defined by coordinating resources.

One consequence of allowing replication operations to target good replicas for
update is that phymv operations can also do so. This is not desired behavior,
so this change ensures that phymv continues to disallow targeting good replicas.

When a client uses keywords to specify a source (either by resource
name or by replica number) AND a destination resource, replication is
expressly disallowed if the resolved hierarchies for source and
destination replicas are the same hierarchy (that is, the source and
destination replicas are the same replica).

When a client only specifies a source (either by resource name or by
replica number) OR a destination resource, replication is allowed even
if the resolved hierarchies for source and destination replicas are the
same hierarchy (that is, the source and destination replicas are the
same replica).

When the source and destination replicas are the same replica, the
replication operation is a no-op, although fileModified will be
triggered so that any configured policy will be in effect. If the
source replica is stale, it cannot be used to update any other
replicas regardless.