[ 7468] Trim: prevent removing the last replica
This commit prevents trim from completely removing a data object by
preventing trim from removing the last replica. The definition of
the meaning of the COPIES_KW (-N) was modified from its original
meaning - "minimum number of good replicas to keep" - to mean
"minimum number of replicas to keep" (notice removal of the word
"good").

Trim now trims down to the specified minimum number of replicas to keep,
also taking stale replicas into consideration. The order of priority
for trimming replicas is such that stale replicas are trimmed first in
order of oldest to newest, and the new replicas are trimmed in order of
oldest to newest. Trim will never remove the last good replica even if
it is unable to trim down to the requested minimum number of replicas.

This commit also changes handling of the COPIES_KW such that providing
any value other than a positive integer results in an error rather
than falling back to the default value of 2.

Also fixes AGE_KW (--age) handling such that it is easier to understand.
The behavior is unchanged.