[ 7476] Replication: DEF_RESC_NAME_KW SOMETIMES acts as a directive
Replication now allows targeting good replicas for update (even though
it results in a no-op). Replication does NOT allow the client to explicitly
specify an identical source and destination replica or resource. Replication
does permit an identical source and destination replica (that is, it does
not result in an error) when the hierarchy resolution logic causes the
situation rather than explicit client directives.

That being said, there is a way to get around the denial of specifying the
same source and destination by specifying a default resource via the
DEF_RESC_NAME_KW. Therefore, replication treats DEF_RESC_NAME_KW the same
as if the client had explicitly specified a destination resource. The one
exception is in the case where the destination resource hierarchy does not
contain the resource indicated by DEF_RESC_NAME_KW. This case is treated as
an error in the case where the client explicitly specified a destination
resource via DEST_RESC_NAME_KW because it was an explicit directive that
could not be serviced by the API. In the case of DEF_RESC_NAME_KW, the
indication of a default resource does not infer that the user DIRECTED the
API to use that resource as the destination resource, but rather that it
is the "preferred" resource. In other words, it should not be considered
an error if the specified default resource does not win the vote.

This is being kept in a separate commit because this is a pretty specific
case and could distract from the more important logic introduced earlier.