[ 7104] Add replica_truncate API
The replica_truncate API is meant to act as a replacement for
rcDataObjTruncate. The interface is very similar to rcDataObjTruncate,
but the result is different. The API only targets one replica to be
truncated, and will trigger fileModified as if the file had been
modified via an open/write/close.