[ 7556] Make find_replica a member function of data_object_proxy
This commit makes irods::experimental::data_object::find_replica a member
function of class template irods::experimental::data_object_proxy, removing
the free functions which required the caller to construct a data_object_proxy
in order to use it. Given that this is in the irods::experimental namespace,
the free functions have been removed.

This commit replaces the find_replica free function variants which have been
removed with a more useful set of functions which take a DataObjInfo&
instead of a data_object_proxy. This allows the caller to take advantage of
the data_object_proxy and replica_proxy interfaces without needing to
construct the objects ahead of time. The returned type is a replica_proxy,
but a pointer to the underlying structure can be obtained through the get()
member function of the replica_proxy object.