[ 2978] GeneralAdmin: Add ability to "add" a "group"
Instead of "add" operation of a "user" with type "rodsgroup",
this commit adds the ability for the "add" operation to
target a "group". This is enforced by returning an error if
the caller attempts to "add" a "user" of type "rodsgroup".

Adds a new test which asserts that SYS_NOT_ALLOWED is returned
on mkuser with a rodsgroup type and removes a series of tests
around using mkuser with rodsgroup types.