[ 8360] pam_password: Do not auto-enable TLS
The legacy PAM authentication plugin and the pam_password
authentication plugin have historically enabled TLS - even if
secure communications are disabled between the client and
server - in order to ensure that PAM passwords are never sent
over the network in the clear.

This commit changes this behavior such that the pam_password
authentication plugin no longer enables TLS and instead will
return an error if TLS is not enabled.