[ 8360] Add test for pam_password with no TLS
This commit also pulls out staticmethods in the single test class
into top-level functions. This was originally because the new test
was going to be in a new class, but it fit well enough in the
existing test class. The staticmethods remain top-level methods
outside of the test class in case we do add more test classes in
the future and these utilities prove useful.