[ 7381] PREP: Add test for CAT_SUCCESS_BUT_WITH_NO_INFO
This commit adds a test which adds a Python rule to core.py which
simply returns CAT_SUCCESS_BUT_WITH_NO_INFO. The rule is then called
from a native rule and ensures that the error is returned (that is,
that it does not get filtered by the PREP).

The test implementation is based on test_python_error_to_native_58.