[ 8441] DataObjUnlink: Unregister before unlinking
This commit swaps the order of unregistering a replica and unlinking
the underlying data in the dataObjUnlinkS function (used by the
DataObjUnlink and DataObjTrim APIs, among others). This is to prevent
accidental data loss caused by eagerly unlinking the data before
checking to see whether the user has sufficient permission to
unregister the data.

This introduces a risk of "leaking" data in a resource should the
unlink fail for some reason. In other words, there would be data in
the vault not registered in the iRODS catalog because the unregister
occurs before the unlink as of this commit. This is acceptable because
ifsck (and other tools) can detect these situations and it is better
than losing the data.